requirements.txt
fastapi==0.115.14
starlette==0.46.2
uvicorn[standard]
SQLAlchemy>=2.0
email-validator>=2.0
asyncpg
pydantic
alembic
python-dotenv
passlib[bcrypt]==1.7.4
python-jose[cryptography]
python-multipart
httpx
aiosqlite
celery
redis
openpyxl
python-telegram-bot==20.3
pywebpush
prometheus-fastapi-instrumentator
aiokafka
clickhouse-connect
pandas
catboost
prophet
mlflow
